refactor: 파일명 형식 조정
