+++ src/components/styles/home.styled.ts
