+++ src/features/admin/AdminLogin.tsx
@@ -5,10 +5,10 @@ import {
   Title,
   Input,
   Button,
-  AdminLoginContainer,
-} from '@/features/user/styles/AdminLoginContainer'
+  AdminLoginStyled,
+} from '@/features/admin/AdminLogin.styled'
 import { loginAdmin } from '@/services/userApi'
-import { ErrorMessage } from '@/features/user/styles/loginStyles'
+import { ErrorMessage } from '@/features/user/styles/login.styled'
 import { colors } from '@/constants'
 
 const AdminLogin = () => {
@@ -33,7 +33,7 @@ const AdminLogin = () => {
   }
 
   return (
-    <AdminLoginContainer>
+    <AdminLoginStyled>
       <Card>
         <Title> 어드민 로그인 </Title>
         <form onSubmit={handleAdminLogin}>
@@ -65,7 +65,7 @@ const AdminLogin = () => {
           </Button>
         </form>
       </Card>
-    </AdminLoginContainer>
+    </AdminLoginStyled>
   )
 }
