+++ src/features/product/styles/controlPanel.styled.ts
