+++ src/features/productDetail/productDetail.types.ts
@@ -4,7 +4,7 @@ export interface SizeInfo {
 }
 
 export interface Product {
-  id: string
+  id: number
   name: string
   price: number
   imageUrl: string
