+++ src/features/user/styles/chooseAvatar.styled.ts
